+++ packages/create/function-templates/p5-video/index.js
@@ -1,6 +1,6 @@
-export const handler = ({ params, mechanic, sketch }) => {
+export const handler = ({ inputs, mechanic, sketch }) => {
   const { width, height, text, color1, color2, radiusPercentage, turns } =
-    params;
+    inputs;
 
   const center = [width / 2, height / 2];
   const radius = ((height / 2) * radiusPercentage) / 100;
@@ -39,7 +39,7 @@ export const handler = ({ params, mechanic, sketch }) => {
   };
 };
 
-export const params = {
+export const inputs = {
   width: {
     type: "number",
     default: 400,