+++ packages/create/function-templates/react-image/index.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 
-export const handler = ({ params, mechanic }) => {
-  const { width, height, text, color1, color2, radiusPercentage } = params;
+export const handler = ({ inputs, mechanic }) => {
+  const { width, height, text, color1, color2, radiusPercentage } = inputs;
 
   const center = [width / 2, height / 2];
   const radius = ((height / 2) * radiusPercentage) / 100;
@@ -42,7 +42,7 @@ export const handler = ({ params, mechanic }) => {
   );
 };
 
-export const params = {
+export const inputs = {
   width: {
     type: "number",
     default: 400,