+++ packages/create/function-templates/react-video/index.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect, useRef } from "react";
 
-export const handler = ({ params, mechanic }) => {
+export const handler = ({ inputs, mechanic }) => {
   const { width, height, text, color1, color2, radiusPercentage, turns } =
-    params;
+    inputs;
 
   const center = [width / 2, height / 2];
   const radius = ((height / 2) * radiusPercentage) / 100;
@@ -52,7 +52,7 @@ export const handler = ({ params, mechanic }) => {
   );
 };
 
-export const params = {
+export const inputs = {
   width: {
     type: "number",
     default: 400,