+++ packages/dsi-logo-maker/functions/textAndImageCanvas/index.js
@@ -6,7 +6,7 @@ import {
 } from "../../utils/blocks";
 import { drawBlock } from "../../utils/blocks-canvas";
 
-export const handler = ({ params, mechanic }) => {
+export const handler = ({ inputs, mechanic }) => {
   const {
     width,
     ratio,
@@ -16,7 +16,7 @@ export const handler = ({ params, mechanic }) => {
     colors: colorsString,
     offset,
     image,
-  } = params;
+  } = inputs;
 
   const words = text.split(" ").map((s) => s.toUpperCase());
   const colors = getColors("Custom Colors", null, colorsString.split(","));
@@ -53,7 +53,7 @@ export const handler = ({ params, mechanic }) => {
   }
 };
 
-export const params = {
+export const inputs = {
   width: {
     type: "number",
     default: 500,