+++ packages/ui-components/src/index.js
@@ -6,5 +6,5 @@ export { TextInput } from "./input/TextInput.js";
 export { NumberInput } from "./input/NumberInput.js";
 export { Select } from "./input/Select.js";
 export { ImageInput } from "./input/ImageInput.js";
-export { ParamInput } from "./ParamInput.js";
+export { MechanicInput } from "./MechanicInput.js";
 export { Dropdown as DropdownIcon, Invalid as InvalidIcon, Add as AddIcon } from "./icons/index.js";